refactor(commands): move to args and utils
